// src/formulas/formulas/contract/xion/types/Treasury.types.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@1.12.1.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export type Addr = string
export type Binary = string
export interface InstantiateMsg {
  admin?: Addr | null
  fee_config: FeeConfig
  grant_configs: GrantConfig[]
  type_urls: string[]
}
export interface FeeConfig {
  allowance?: Any | null
  description: string
  expiration?: number | null
}
export interface Any {
  type_url: string
  value: Binary
}
export interface GrantConfig {
  authorization: Any
  description: string
  optional: boolean
}
export type ExecuteMsg =
  | {
      propose_admin: {
        new_admin: Addr
      }
    }
  | {
      accept_admin: {}
    }
  | {
      cancel_proposed_admin: {}
    }
  | {
      update_grant_config: {
        grant_config: GrantConfig
        msg_type_url: string
      }
    }
  | {
      remove_grant_config: {
        msg_type_url: string
      }
    }
  | {
      update_fee_config: {
        fee_config: FeeConfig
      }
    }
  | {
      deploy_fee_grant: {
        authz_grantee: Addr
        authz_granter: Addr
      }
    }
  | {
      revoke_allowance: {
        grantee: Addr
      }
    }
  | {
      update_params: {
        params: Params
      }
    }
  | {
      withdraw: {
        coins: Coin[]
      }
    }
  | {
      migrate: {
        migrate_msg: Binary
        new_code_id: number
      }
    }
export type Uint128 = string
export interface Params {
  display_url: string
  redirect_url: string
  icon_url: string
}
export interface Coin {
  amount: Uint128
  denom: string
}
export type QueryMsg =
  | {
      grant_config_by_type_url: {
        msg_type_url: string
      }
    }
  | {
      grant_config_type_urls: {}
    }
  | {
      fee_config: {}
    }
  | {
      admin: {}
    }
  | {
      pending_admin: {}
    }
  | {
      params: {}
    }
export type NullableAddr = Addr | null
export type ArrayOfString = string[]